add SettingForm component
